requirements.txt
fastapi==0.75.0
uvicorn[standard]==0.17.0
pydantic[dotenv]
yarl==1.7.2
ujson==5.1.0
psycopg2-binary
sqlalchemy==1.4.39
alembic==1.7.7
redis[hiredis]==4.3.3
httptools==0.3.0
loguru==0.6.0
pytest==7.0
flake8==4.0.1
mypy==0.961
isort==5.10.1
yesqa==1.3.0
pre-commit==2.19.0
wemake-python-styleguide==0.16.1
black==22.3.0
autoflake==1.4
pytest-cov==3.0.0
anyio==3.6.1
pytest-env==0.6.2
fakeredis==1.8.1
httpx
pyrebase4 
firebase-admin